Re-added delimiters and added some more helper functions
Added in the delimiters again because the issues with quotes was screwing up the code yesterday, and I see now why they are useful. Added in functions that convert the result lists into dictionaries giving an ID to each agent. Added a function that converts the level1 results to their preferences.