Re-implement randomisation
Removed the new_ prefix from the refactored functions.
Added a default argument randomisation=False to the function which generates the ID for the agents. If randomisation is set to True, the whole algorithm will run slightly differently each time.
Added a default argument first_line=True to the function which pulls the data from the csv files into a list. This stops the column headings appearing which makes things nicer.